refactor(app): flatten runtime state into app state
