+++ src/app/cmd/er/handler.rs
@@ -228,7 +228,7 @@ async fn handle_write_failure_log(
     state: &AppState,
     failed_tables: Vec<(String, String)>,
 ) -> Result<()> {
-    match config_writer.get_cache_dir(state.runtime.project_name()) {
+    match config_writer.get_cache_dir(state.project_name()) {
         Ok(cache_dir) => {
             let writer = Arc::clone(er_log_writer);
             let tx = action_tx.clone();
