+++ src/app/model/mod.rs
@@ -6,6 +6,5 @@ pub mod shared;
 pub mod sql_editor;
 
 pub mod app_state;
-pub mod runtime_state;
 pub mod sqlite;
 pub mod table_prefetch;
